Implement service metrics and route logging in the CLI and engine
- Added a new `list_app_routes` function to retrieve and sort application routes for logging.
- Enhanced the CLI to log available routes and server address upon startup.
- Introduced metrics logging capabilities in the `AnnaEngine` to track request statistics, including total requests started, completed, and failed.
- Updated `ServeSettings` to include a new configuration option for metrics logging interval.
- Added tests to validate the new route logging and service metrics functionality.

+++ tests/test_fused_op_xpu.py
@@ -18,7 +18,7 @@ def test_gated_delta_fused_xpu_matches_reference() -> None:
     g = torch.randn(2, 3, 4, device=device, dtype=torch.float32)
     beta = torch.sigmoid(torch.randn(2, 3, 4, device=device, dtype=torch.float16))
     z = torch.randn(2, 3, 4, 8, device=device, dtype=torch.float16)
-    norm_weight = torch.randn(8, device=device, dtype=torch.float16)
+    norm_weight = torch.randn(8, device=device, dtype=torch.float32)
     initial_state = torch.randn(2, 4, 8, 8, device=device, dtype=torch.float32)
 
     output, final_state = torch.ops.anna.gated_delta_fused(
